requirements.txt
discord.py
openai==0.28
python-dotenv
loguru
